Restructure the modules
