+++ src/modules/layout/sidebar.js
@@ -1,12 +1,15 @@
-import projectList from './project-list';
+import projectList from '../objects/project-list';
+import activeProject from '../active-project';
 
 const sidebarLinks = () => {
   const linkContainer = document.createElement('div');
   linkContainer.setAttribute('id', 'link-container');
 
-  projectList.forEach(project => {
+  projectList.forEach((project, index) => {
       const link = document.createElement('a');
-
+      link.onclick = () => {
+        activeProject(index);
+      }
       const linkTitle = document.createElement('p');
       linkTitle.textContent += project.title;
 
@@ -19,10 +22,11 @@ const sidebarLinks = () => {
 const sideBar = () => {
     const sideNavigation = document.createElement('div');
     sideNavigation.classList.add('side-bar');
+    sideNavigation.setAttribute('id', 'side-bar');
 
     sideNavigation.appendChild(sidebarLinks());
 
     return sideNavigation;
 }
 
-export default sideBar;
+export { sideBar, sidebarLinks };
