+++ src/modules/objects/project-class.js
